+++ integration/css-side-effect-imports-test.ts
@@ -25,7 +25,6 @@ test.describe("CSS side-effect imports", () => {
             future: {
               // Enable all CSS future flags to
               // ensure features don't clash
-              unstable_cssModules: true,
               unstable_cssSideEffectImports: true,
               unstable_postcss: true,
               unstable_tailwind: true,