+++ integration/defer-test.ts
@@ -71,11 +71,9 @@ test.describe("non-aborted", () => {
           import Counter from "~/components/counter";
           import Interactive from "~/components/interactive";
 
-          export const meta: MetaFunction = () => ({
-            charset: "utf-8",
-            title: "New Remix App",
-            viewport: "width=device-width,initial-scale=1",
-          });
+          export function meta() {
+            return [{ title: "New Remix App" }];
+          }
 
           export const loader = () => defer({
             id: "${ROOT_ID}",
@@ -86,6 +84,8 @@ test.describe("non-aborted", () => {
             return (
               <html lang="en">
                 <head>
+                  <meta charSet="utf-8" />
+                  <meta name="viewport" content="width=device-width,initial-scale=1" />
                   <Meta />
                   <Links />
                 </head>
@@ -1100,11 +1100,9 @@ test.describe("aborted", () => {
           import Counter from "~/components/counter";
           import Interactive from "~/components/interactive";
 
-          export const meta: MetaFunction = () => ({
-            charset: "utf-8",
-            title: "New Remix App",
-            viewport: "width=device-width,initial-scale=1",
-          });
+          export function meta() {
+            return [{ title: "New Remix App" }];
+          }
 
           export const loader = () => defer({
             id: "${ROOT_ID}",
@@ -1115,6 +1113,8 @@ test.describe("aborted", () => {
             return (
               <html lang="en">
                 <head>
+                  <meta charSet="utf-8" />
+                  <meta name="viewport" content="width=device-width,initial-scale=1" />
                   <Meta />
                   <Links />
                 </head>