+++ packages/remix-react/index.tsx
@@ -68,10 +68,9 @@ export type { ThrownResponse } from "./errors";
 
 export type { HtmlLinkDescriptor } from "./links";
 export type {
-  HtmlMetaDescriptor,
-  V2_MetaArgs,
-  V2_MetaDescriptor,
-  V2_MetaFunction,
+  MetaArgs,
+  MetaDescriptor,
+  MetaFunction,
   RouteModules as UNSAFE_RouteModules,
 } from "./routeModules";