+++ packages/remix-server-runtime/entry.ts
@@ -23,13 +23,11 @@ type VanillaExtractOptions = {
 };
 
 export interface FutureConfig {
-  unstable_cssModules: boolean;
   unstable_cssSideEffectImports: boolean;
   unstable_dev: boolean | Dev;
   unstable_postcss: boolean;
   unstable_tailwind: boolean;
   unstable_vanillaExtract: boolean | VanillaExtractOptions;
-  v2_meta: boolean;
   v2_routeConvention: boolean;
 }