+++ packages/remix-server-runtime/reexport.ts
@@ -39,17 +39,14 @@ export type {
   DataFunctionArgs,
   ErrorBoundaryComponent,
   HeadersFunction,
-  HtmlMetaDescriptor,
   LinksFunction,
   LoaderArgs,
   LoaderFunction,
-  MetaDescriptor,
-  MetaFunction,
   RouteComponent,
   RouteHandle,
-  V2_ServerRuntimeMetaArgs,
-  V2_ServerRuntimeMetaDescriptor,
-  V2_ServerRuntimeMetaFunction,
+  ServerRuntimeMetaArgs,
+  ServerRuntimeMetaDescriptor,
+  ServerRuntimeMetaFunction,
 } from "./routeModules";
 
 export type { SerializeFrom } from "./serialize";