+++ scripts/playground/template/app/routes/join.tsx
@@ -64,9 +64,7 @@ export async function action({ request }: ActionArgs) {
 }
 
 export const meta: MetaFunction = () => {
-  return {
-    title: "Sign Up",
-  };
+  return [{ title: "Sign Up" }];
 };
 
 export default function Join() {