+++ scripts/playground/template/app/routes/notes/index.tsx
@@ -1,4 +1,9 @@
+import type { MetaFunction } from "@remix-run/node";
 import { Link } from "@remix-run/react";
+
+export const meta: MetaFunction = () => {
+  return [{ title: "Remix Notes" }];
+};
 
 export default function NoteIndexPage() {
   return (